refactor: move pattern validation for datagrabbers to PatternValidationMixin
